refactor: move RentalIntervalStateModifier to its own file
