+++ script-rental/src/main/kotlin/com/cereal/rental/RentalIntervalStateModifier.kt
@@ -1,0 +1,20 @@
+package com.cereal.rental
+
+import com.cereal.script.utils.configuration.BaseConfiguration
+import com.cereal.sdk.statemodifier.ScriptConfig
+import com.cereal.sdk.statemodifier.ScriptConfigValue
+import com.cereal.sdk.statemodifier.StateModifier
+import com.cereal.sdk.statemodifier.Visibility
+
+object RentalIntervalStateModifier : StateModifier {
+    override fun getError(scriptConfig: ScriptConfig): String? {
+        val value = scriptConfig.valueForKey(BaseConfiguration.KEY_MONITOR_INTERVAL)
+        return if (value is ScriptConfigValue.IntScriptConfigValue && value.value < 1) {
+            "Interval must be at least 1 minute."
+        } else {
+            null
+        }
+    }
+
+    override fun getVisibility(scriptConfig: ScriptConfig): Visibility = Visibility.VisibleOptional
+}
